added logic to driver class
