+++ WastePermitsAutomation/Pages/MainPage.cs
