+++ WastePermitsAutomation/WastePermitsAutomation.csproj
@@ -49,10 +49,10 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="DashboardPage.cs" />
-    <Compile Include="Driver.cs" />
-    <Compile Include="LoginPage.cs" />
-    <Compile Include="MainPage.cs" />
+    <Compile Include="Pages\DashboardPage.cs" />
+    <Compile Include="Selenium\Driver.cs" />
+    <Compile Include="Pages\LoginPage.cs" />
+    <Compile Include="Pages\MainPage.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
