+++ WastePermitsTests/WastePermitsTest.cs
@@ -1,0 +1,23 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using WastePermitsAutomation;
+
+namespace WastePermitsTests
+{
+    public class WastePermitsTest
+    {
+        [TestInitialize]
+        public void Init()
+        {
+            Driver.Initialize();
+            LoginPage.Goto();
+            LoginPage.LoginAs("tim.stone@defradev.onmicrosoft.com").WithPassword("Helsink!1").Login();
+        }
+
+        [TestCleanup]
+        public void Cleanup()
+        {
+            Driver.Close();
+        }
+
+    }
+}
