+++ WastePermitsTests/WastePermitsTests.csproj
@@ -51,6 +51,7 @@
   <ItemGroup>
     <Compile Include="LoginTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="WastePermitsTest.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
